+++ src/modules/tools/statisticDashboard/utils/handleFeatures.js
@@ -15,7 +15,8 @@ function filterFeaturesByKeyValue (features, key, value) {
         return [];
     }
 
-    return features.filter(feature => feature.get(key).split(value).length - 1 === 1);
+    return features.filter(feature => typeof feature.get(key) !== "undefined"
+        && feature.get(key).split(value).length - 1 === 1);
 }
 
 /**